ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
huggingface_hub
requests
pydantic
